backend/requirements-dev.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
six==1.15.0
virtualenv==20.4.2
Django==3.1.7
djangorestframework==3.12.2
drf-spectacular==0.13.2
Pillow==8.1.2
psycopg2-binary==2.8.6
jsonschema==3.2.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
virtualenv==20.4.2
gunicorn~=20.0
s3transfer==0.3.7
boto3==1.17.53
botocore==1.20.53
django-storages==1.11.1

